+++ backend/ormconfig.json
@@ -5,9 +5,9 @@
   "username": "root",
   "password": "",
   "database": "gobarber",
-  "entities": [ "./src/models/*.ts" ],
-  "migrations": [ "./src/database/migrations/*.ts" ],
+  "entities": ["./src/modules/**/infra/typeorm/entities/*.ts"],
+  "migrations": [ "./src/shared/infra/typeorm/migrations/*.ts" ],
   "cli": {
-    "migrationsDir": "./src/database/migrations"
+    "migrationsDir": "./src/shared/infra/typeorm/migrations"
   }
 }
