+++ backend/src/modules/appointments/infra/http/routes/appointments.routes.ts
@@ -2,9 +2,9 @@ import { Router } from 'express'
 import { parseISO } from 'date-fns'
 import { getCustomRepository } from 'typeorm'
 
-import AppointmentsRepository from '../repositories/AppointmentsRepository'
-import CreateAppointmentService from '../services/CreateAppointmentService'
-import ensureAuthenticated from '../middlewares/ensureAuthenticated'
+import AppointmentsRepository from '@/modules/appointments/repositories/AppointmentsRepository'
+import CreateAppointmentService from '@/modules/appointments/services/CreateAppointmentService'
+import ensureAuthenticated from '@/modules/users/infra/http/middlewares/ensureAuthenticated'
 
 const appointmentsRouter = Router()
