+++ backend/src/modules/appointments/repositories/AppointmentsRepository.ts
@@ -1,6 +1,6 @@
 import { EntityRepository, Repository } from 'typeorm'
 
-import Appointment from '../models/Appointment'
+import Appointment from '@/modules/appointments/infra/typeorm/entities/Appointment'
 
 @EntityRepository(Appointment)
 export default class AppointmentsRepository extends Repository<Appointment> {
