+++ backend/src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,6 +1,6 @@
 import { Router } from 'express'
 
-import AuthenticateUserService from '../services/AuthenticateUserService'
+import AuthenticateUserService from '@/modules/users/services/AuthenticateUserService'
 
 const sessionsRouter = Router()
