+++ backend/src/shared/infra/http/routes/index.ts
@@ -1,0 +1,13 @@
+import { Router } from 'express'
+
+import appointmentsRouter from '@/modules/appointments/infra/http/routes/appointments.routes'
+import sessionsRouter from '@/modules/users/infra/http/routes/sessions.routes'
+import usersRouter from '@/modules/users/infra/http/routes/users.routes'
+
+const routes = Router()
+
+routes.use('/appointments', appointmentsRouter)
+routes.use('/sessions', sessionsRouter)
+routes.use('/users', usersRouter)
+
+export default routes
