+++ backend/src/shared/infra/http/server.ts
@@ -2,10 +2,10 @@ import 'reflect-metadata'
 import express, { Request, Response, NextFunction } from 'express'
 import 'express-async-errors'
 
-import routes from './routes'
-import './database'
-import { tmpDirectory } from './config/upload'
-import AppError from './errors/AppError'
+import { tmpDirectory } from '@/config/upload'
+import AppError from '@/shared/errors/AppError'
+import routes from '@/shared/infra/http/routes'
+import '@/shared/infra/typeorm'
 
 const app = express()
